[Fixes #9] Add cross-platform camera support and tighten data connection handling (#193)
* **New Features**
  * Added camera support with live JPEG streaming for desktop (OpenCV) and ESP32 platforms.
  * Introduced a CameraWorkload to capture and share camera frames.
  * Web interface now displays a live camera feed, fetching images periodically.

* **Improvements**
  * Enforced stricter validation for workload data connection paths and uniqueness.
  * Converted most timing and numeric fields from double to float for better performance and consistency.
  * Enhanced web server to handle large binary responses using fixed-size buffers.

* **Bug Fixes**
  * Generalized data connection and field traversal logic for improved flexibility and reliability.
  * Removed unnecessary includes and fixed packet logging behavior.

* **Documentation**
  * Updated design documents and code comments to reflect new data types and behaviors.

* **Chores**
  * Added OpenCV dependency for desktop builds; updated setup scripts, CMake, and Docker configurations accordingly.
  * Improved CMake symlink creation with error handling.

* **Style**
  * Refined CSS and HTML for improved camera feed integration and responsive layout.
  * Updated code formatting and naming conventions for clarity.

* **Tests**
  * Expanded tests to cover new data types, connection rules, and camera functionality.
  * Adjusted tests to reflect stricter connection path semantics and float usage.

+++ cpp/include/robotick/framework/registry/WorkloadTypeHelpers.h
@@ -58,7 +58,7 @@ namespace robotick::registry
 	template <typename T, typename = void> struct has_start : std::false_type
 	{
 	};
-	template <typename T> struct has_start<T, std::void_t<decltype(std::declval<T>().start(std::declval<double>()))>> : std::true_type
+	template <typename T> struct has_start<T, std::void_t<decltype(std::declval<T>().start(std::declval<float>()))>> : std::true_type
 	{
 	};
 
@@ -168,7 +168,7 @@ namespace robotick::registry
 		static_cast<T*>(self)->setup();
 	}
 
-	template <typename T> static void start_fn(void* self, double t)
+	template <typename T> static void start_fn(void* self, float t)
 	{
 		static_cast<T*>(self)->start(t);
 	}
