+++ cpp/include/robotick/framework/utils/ConsoleTelemetryTable.h
@@ -16,15 +16,27 @@ namespace robotick
 		std::string config;
 		std::string inputs;
 		std::string outputs;
-		double tick_duration_ms = 0.0; // actual work time
-		double tick_delta_ms = 0.0;	   // time since last tick call
-		double goal_interval_ms = 0.0; // expected interval
+		float tick_duration_ms = 0.0; // actual work time
+		float tick_delta_ms = 0.0;	  // time since last tick call
+		float goal_interval_ms = 0.0; // expected interval
 
 		ConsoleTelemetryRow() = default;
-		ConsoleTelemetryRow(std::string type, std::string name, std::string config, std::string inputs, std::string outputs, double tick_duration_ms,
-			double tick_delta_ms, double goal_interval_ms)
-			: type(std::move(type)), name(std::move(name)), config(std::move(config)), inputs(std::move(inputs)), outputs(std::move(outputs)),
-			  tick_duration_ms(tick_duration_ms), tick_delta_ms(tick_delta_ms), goal_interval_ms(goal_interval_ms)
+		ConsoleTelemetryRow(std::string type,
+			std::string name,
+			std::string config,
+			std::string inputs,
+			std::string outputs,
+			float tick_duration_ms,
+			float tick_delta_ms,
+			float goal_interval_ms)
+			: type(std::move(type))
+			, name(std::move(name))
+			, config(std::move(config))
+			, inputs(std::move(inputs))
+			, outputs(std::move(outputs))
+			, tick_duration_ms(tick_duration_ms)
+			, tick_delta_ms(tick_delta_ms)
+			, goal_interval_ms(goal_interval_ms)
 		{
 		}
 	};
