+++ cpp/src/robotick/framework/Engine.cpp
@@ -274,7 +274,7 @@ namespace robotick
 
 		state->is_running = true;
 
-		const auto child_tick_interval_sec = std::chrono::duration<double>(1.0 / root_tick_rate_hz);
+		const auto child_tick_interval_sec = std::chrono::duration<float>(1.0f / root_tick_rate_hz);
 		const auto child_tick_interval = std::chrono::duration_cast<std::chrono::steady_clock::duration>(child_tick_interval_sec);
 
 		const auto engine_start_time = std::chrono::steady_clock::now() - child_tick_interval;
