+++ cpp/src/robotick/workloads/diagnostics/TimingDiagnosticsWorkload.cpp
@@ -28,14 +28,14 @@ namespace robotick
 
 	struct TimingDiagnosticsOutputs
 	{
-		double last_tick_rate = 0.0;
-		double avg_tick_rate = 0.0;
-		double tick_stddev = 0.0;
+		float last_tick_rate = 0.0;
+		float avg_tick_rate = 0.0;
+		float tick_stddev = 0.0;
 	};
 	ROBOTICK_REGISTER_STRUCT_BEGIN(TimingDiagnosticsOutputs)
-	ROBOTICK_STRUCT_FIELD(TimingDiagnosticsOutputs, double, last_tick_rate)
-	ROBOTICK_STRUCT_FIELD(TimingDiagnosticsOutputs, double, avg_tick_rate)
-	ROBOTICK_STRUCT_FIELD(TimingDiagnosticsOutputs, double, tick_stddev)
+	ROBOTICK_STRUCT_FIELD(TimingDiagnosticsOutputs, float, last_tick_rate)
+	ROBOTICK_STRUCT_FIELD(TimingDiagnosticsOutputs, float, avg_tick_rate)
+	ROBOTICK_STRUCT_FIELD(TimingDiagnosticsOutputs, float, tick_stddev)
 	ROBOTICK_REGISTER_STRUCT_END(TimingDiagnosticsOutputs)
 
 	// === Internal state (not registered) ===
@@ -43,8 +43,8 @@ namespace robotick
 	struct TimingDiagnosticsInternalState
 	{
 		int count = 0;
-		double sum_dt = 0.0;
-		double sum_dt2 = 0.0;
+		float sum_dt = 0.0;
+		float sum_dt2 = 0.0;
 	};
 
 	// === Workload ===
@@ -66,13 +66,13 @@ namespace robotick
 
 		void tick(const TickInfo& tick_info)
 		{
-			if (config.log_rate_hz == 0 || tick_info.delta_time <= 0.0)
+			if (config.log_rate_hz == 0 || tick_info.delta_time <= 0.0f)
 			{
 				return;
 			}
 
-			const double actual_dt = tick_info.delta_time;
-			const double tick_rate = 1.0 / actual_dt;
+			const float actual_dt = tick_info.delta_time;
+			const float tick_rate = 1.0f / actual_dt;
 
 			outputs.last_tick_rate = tick_rate;
 
@@ -80,27 +80,27 @@ namespace robotick
 			internal_state.sum_dt += actual_dt;
 			internal_state.sum_dt2 += actual_dt * actual_dt;
 
-			const double tick_period = 1.0 / config.log_rate_hz;
+			const float tick_period = 1.0f / (float)config.log_rate_hz;
 
 			if (internal_state.sum_dt >= tick_period)
 			{
-				double mean_dt = internal_state.sum_dt / internal_state.count;
-				double mean_dt2 = internal_state.sum_dt2 / internal_state.count;
-				const double variance = std::max(0.0, mean_dt2 - mean_dt * mean_dt);
-				const double stddev = std::sqrt(variance);
+				float mean_dt = internal_state.sum_dt / (float)internal_state.count;
+				float mean_dt2 = internal_state.sum_dt2 / (float)internal_state.count;
+				const float variance = std::max(0.0f, mean_dt2 - mean_dt * mean_dt);
+				const float stddev = std::sqrt(variance);
 
-				outputs.avg_tick_rate = 1.0 / mean_dt;
+				outputs.avg_tick_rate = 1.0f / mean_dt;
 				outputs.tick_stddev = stddev;
 
-				static constexpr double seconds_to_microseconds = 1e6;
+				static constexpr float seconds_to_microseconds = 1e6f;
 
 				std::cerr << std::fixed;
 				std::cerr << "[TimingDiagnostics] avg: " << outputs.avg_tick_rate << " Hz, stddev: " << outputs.tick_stddev * seconds_to_microseconds
 						  << " µs\n";
 
 				internal_state.count = 0;
-				internal_state.sum_dt = 0.0;
-				internal_state.sum_dt2 = 0.0;
+				internal_state.sum_dt = 0.0f;
+				internal_state.sum_dt2 = 0.0f;
 			}
 		}
 	};
